requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --no-annotate --output-file requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6 ; sys_platform == 'win32'
distro==1.9.0
gitdb==4.0.12
gitpython==3.1.46
google-auth==2.47.0
google-genai==1.58.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
narwhals==2.15.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.5.2
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==1.26.20
watchdog==6.0.0 ; sys_platform != 'darwin'
websockets==15.0.1
